requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
matplotlib>=3.4.3
scikit-learn>=0.24.2 